llm_connector_service/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
anthropic

sentence-transformers
faiss-cpu
